requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# Google APIs
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Vector Database and ML
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3

# Database
sqlite3

# OpenAI
openai==1.3.7

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1

# Environment Variables
python-dotenv==1.0.0

# Docker (for development)
docker==6.1.3

# Additional utilities
pydantic==2.5.0
asyncio-mqtt==0.16.1
aiofiles==23.2.1
schedule==1.2.0
pytz==2023.3

# Logging
loguru==0.7.2

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
